ozbalerts/requirements.txt
slackclient==2.7.2
pyyaml==3.13
requests==2.22.0